type change (int -> size_t)
Former-commit-id: edca09c5886638249d1e95cc0cb1b97129295a33

+++ src/inscore/signal/GraphicSignal.cpp
@@ -85,7 +85,7 @@ GraphicFrames::GraphicFrames(const vector<float>& value)
 	: Frame (value)
 {
 	fSize = int(ceil (float(value.size()) / GraphicFrame::FrameSize()));
-	int fill = fSize * GraphicFrame::FrameSize();
+	size_t fill = fSize * GraphicFrame::FrameSize();
 	for (size_t i = value.size(); i < fill; i++)
 		fData.push_back(0.f);
 }
